[hello-world] Adds test
