+++ seeds/hello-world/src/board.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright 2023 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+import { Board } from "@google-labs/breadboard";
+
+const board = new Board();
+const input = board.input();
+const output = board.output();
+
+input.wire("say->hear", output);
+export const result = await board.runOnce({
+  say: "Hello, world?",
+});
