+++ seeds/hello-world/src/index.ts
@@ -3,14 +3,6 @@
  * Copyright 2023 Google LLC
  * SPDX-License-Identifier: Apache-2.0
  */
-import { Board } from "@google-labs/breadboard";
-const board = new Board();
-const input = board.input();
-const output = board.output();
+import { result } from "./board.js";
 
-input.wire("say->hear", output);
-const result = await board.runOnce({
-  say: "Hello, world?",
-});
-
-console.log("result", result);
+console.log(result);
